+++ src/Moongate.Persistence/Data/Internal/PersistenceStateStore.cs
@@ -21,9 +21,9 @@ internal sealed class PersistenceStateStore
 
     public long LastSequenceId { get; set; }
 
-    public uint LastAccountId { get; set; } = (uint)(Serial.MobileStart - 1);
+    public uint LastAccountId { get; set; } = Serial.MobileStart - 1;
 
-    public uint LastMobileId { get; set; } = (uint)(Serial.MobileStart - 1);
+    public uint LastMobileId { get; set; } = Serial.MobileStart - 1;
 
     public uint LastItemId { get; set; } = Serial.ItemOffset - 1;
 }
